fix: [Adapter] DeploymentId is not automatically inserted into the model endpoint field
