+++ apps/ai-dial-admin/src/app/[lang]/models/actions.ts
@@ -8,6 +8,8 @@ import { DialModel, DialModelType } from '@/src/models/dial/model';
 import { getUserToken } from '@/src/utils/auth/auth-request';
 import { getIsEnableAuthToggle } from '@/src/utils/env/get-auth-toggle';
 import { convertDefaultsToRecord } from '@/src/components/Defaults/utils';
+import { SOURCE_FIELD } from '@/src/components/SourceField/types';
+import { getEndpointPostfix } from '@/src/utils/models/model-endpoint';
 
 export async function getModels() {
   const token = await getUserToken(getIsEnableAuthToggle(), headers(), cookies());
@@ -52,6 +54,10 @@ export async function createModel(model: DialModel) {
       ...model,
       ...DEFAULT_ROLE_LIMITS,
       type: model.type || DialModelType.Chat,
+      source: {
+        ...model.source,
+        completionEndpointPath: `${model.name}${getEndpointPostfix(DialModelType.Chat)}`,
+      } as SOURCE_FIELD,
     },
     token,
   );
