+++ apps/ai-dial-admin/src/utils/models/model-endpoint.ts
@@ -1,0 +1,5 @@
+import { DialModelType } from '@/src/models/dial/model';
+
+export const getEndpointPostfix = (type?: DialModelType) => {
+  return type === DialModelType.Chat ? '/chat/completions' : '/embeddings';
+};
